+++ WhatsAppAPISolutionDL/Models/Media.cs
@@ -5,7 +5,7 @@ namespace WhatsAppAPISolutionDL.Models
 {
     public partial class Media
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public int? ClientId { get; set; }
         public string WhatsAppBusinessAccountId { get; set; }
         public int? SenderNameId { get; set; }
@@ -14,7 +14,7 @@ namespace WhatsAppAPISolutionDL.Models
         public string MediaPath { get; set; }
         public int? RecordStatus { get; set; }
         public string ContentType { get; set; }
-        public long? FileSize { get; set; }
+        public int? FileSize { get; set; }
         public string FileName { get; set; }
         public string FileExtension { get; set; }
     }
