+++ WhatsAppAPISolutionDL/Models/UsersRole.cs
@@ -5,10 +5,10 @@ namespace WhatsAppAPISolutionDL.Models
 {
     public partial class UsersRole
     {
-        public long UserRoleId { get; set; }
+        public int UserRoleId { get; set; }
         public int? ClientId { get; set; }
-        public long? UserId { get; set; }
-        public long? RoleId { get; set; }
+        public int? UserId { get; set; }
+        public int? RoleId { get; set; }
         public string CreatedBy { get; set; }
         public DateTime? CreatedDate { get; set; }
         public int? UpdatedBy { get; set; }
