+++ WhatsAppAPISolutionDL/UserModels/UClient.cs
@@ -8,7 +8,7 @@ namespace WhatsAppAPISolutionDL.UserModels
 {
     public partial class UClient
     {
-        public long ClientId { get; set; }
+        public int ClientId { get; set; }
         public string ClientName { get; set; }
         public int? ClientLanguage { get; set; }
         public string ClientAddress { get; set; }
