+++ WhatsAppAPISolutionDL/UserModels/UMediaUpload.cs
@@ -8,7 +8,7 @@ namespace WhatsAppAPISolutionDL.UserModels
 {
     public partial class UMediaUpload
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public int? ClientId { get; set; }
         public string WhatsAppBusinessAccountId { get; set; }
         public int? SenderNameId { get; set; }
@@ -16,7 +16,7 @@ namespace WhatsAppAPISolutionDL.UserModels
         public string MediaId { get; set; }
         public string MediaPath { get; set; }
         public string ContentType { get; set; }
-        public long? FileSize { get; set; }
+        public int? FileSize { get; set; }
         public string FileName { get; set; }
         public string FileExtension { get; set; }
     }
